+++ examples/kubernetes/cluster/standard/main.tf
@@ -1,5 +1,6 @@
 resource "vkcs_kubernetes_cluster" "k8s-cluster" {
-  name                = "k8s-cluster"
+  name                = "k8s-standard-cluster"
+  cluster_type        = "standard"
   cluster_template_id = data.vkcs_kubernetes_clustertemplate.k8s_24.id
   master_flavor       = data.vkcs_compute_flavor.basic.id
   master_count        = 1